feat: add create_landmarks_in_service function for landmark snapshot creation
